+++ crates/teloxide-core/src/adaptors/throttle/requester_impl.rs
@@ -200,6 +200,7 @@ where
         answer_pre_checkout_query,
         get_star_transactions,
         refund_star_payment,
+        edit_user_star_subscription,
         set_passport_data_errors,
         set_game_score,
         set_game_score_inline,